# traffic-safety-openapi/public/api/docs/openapi.json
openapi: 3.0.0
info:
  title: 'Traffic Safety Platform API'
  description: 'API documentation for the Traffic Safety decision management system.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'Relative base URL'
paths:
  /api/accidents:
    get:
      tags:
        - Accidents
      summary: 'List all recorded accidents.'
      operationId: listAccidents
      responses:
        '200':
          description: 'Collection of accidents.'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Accident' } }
                type: object
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Accidents
      summary: 'Create a new accident record.'
      operationId: createAccident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: 'Accident payload as accepted by the AccidentFactory.'
              properties:
                id:
                  type: integer
                  format: int64
              type: object
              additionalProperties: true
      responses:
        '201':
          description: 'Accident created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accident'
        '422':
          description: 'Validation error.'
      security:
        -
          bearerAuth: []
  '/api/accidents/{id}':
    get:
      tags:
        - Accidents
      summary: 'Retrieve a single accident by its identifier.'
      operationId: getAccident
      parameters:
        -
          name: id
          in: path
          description: 'Accident identifier'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Accident found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accident'
        '404':
          description: 'Accident not found.'
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Accidents
      summary: 'Update an existing accident record.'
      operationId: updateAccident
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  type: integer
                  format: int64
                  nullable: true
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 'Accident updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accident'
        '404':
          description: 'Accident not found.'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Accidents
      summary: 'Delete an accident.'
      operationId: deleteAccident
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'Accident deleted.'
      security:
        -
          bearerAuth: []
  /api/accidents/total-estimated-cost:
    get:
      tags:
        - Accidents
      summary: 'Get the total estimated cost for all accidents.'
      operationId: getTotalEstimatedAccidentCost
      responses:
        '200':
          description: 'Aggregate cost.'
          content:
            application/json:
              schema:
                properties:
                  totalEstimatedCost: { type: number, format: double }
                type: object
      security:
        -
          bearerAuth: []
  /api/accidents/calculate-total-cost:
    post:
      tags:
        - Accidents
      summary: 'Calculate the combined cost for provided accidents payload.'
      operationId: calculateAccidentTotalCost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - accidents
              properties:
                accidents:
                  type: array
                  items: { type: object, additionalProperties: true }
              type: object
      responses:
        '200':
          description: 'Total cost calculated.'
          content:
            application/json:
              schema:
                properties:
                  totalCost: { type: number, format: double }
                type: object
        '422':
          description: 'Invalid payload.'
      security:
        -
          bearerAuth: []
  /api/accidents/search:
    post:
      tags:
        - Accidents
      summary: 'Search accidents using filters.'
      operationId: searchAccidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: 'Search filters as defined by AccidentSearchDTO.'
              type: object
      responses:
        '200':
          description: 'Search results.'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Accident' } }
                  count: { type: integer }
                type: object
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    MonetaryAmount:
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD
      type: object
...
    AccidentLocation:
      required:
        - locationType
        - locationId
      properties:
        locationType:
          type: string
          example: road_segment
        locationId:
          type: integer
          format: int64
        latitude:
          type: number
          format: float
          nullable: true
        longitude:
          type: number
          format: float
          nullable: true
        distanceFromStart:
          type: number
          format: float
          nullable: true
      type: object
    Accident:
      required:
        - id
        - occurredAt
        - location
        - type
      properties:
        id:
          type: integer
          format: int64
        occurredAt:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/AccidentLocation'
        cost:
          oneOf:
            -
              $ref: '#/components/schemas/MonetaryAmount'
          nullable: true
        type:
          type: string
        severity:
          type: string
          nullable: true
        collisionType:
          type: string
          nullable: true
        causeFactor:
          type: string
          nullable: true
        weatherCondition:
          type: string
          nullable: true
        roadCondition:
          type: string
          nullable: true
        visibilityCondition:
          type: string
          nullable: true
        injuredPersonsCount:
          type: integer
          nullable: true
        locationDescription:
          type: string
          nullable: true
        requiresImmediateAttention:
          type: boolean
        daysSinceOccurrence:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...